moved the services to the folder service
