+++ GoTripApplication/src/app/services/getCurrentUserData.service.ts
