+++ GoTripApplication/src/app/services/getTrip.service.ts
